requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
openai==0.27.0
python-dotenv==0.19.0
httpx==0.24.1 